Move model and icon-loaders to resources.util
