+++ miniml/fileio/util.py
@@ -7,3 +7,11 @@ def is_keras_model(filepath: str) -> bool:
     """
     with h5py.File(filepath, "r") as f:
         return "keras_version" in f.attrs
+
+
+def get_hdf_keys(filepath: str) -> list:
+    """
+    Returns a list of keys in an hdf5 file.
+    """
+    with h5py.File(filepath, "r") as f:
+        return list(f.keys())
