+++ miniml/gui/views/event_viewer.py
@@ -13,7 +13,7 @@ from PyQt5.QtWidgets import (
     QToolBar,
 )
 
-from miniml.gui.util import get_icon_file_path
+from miniml.resources.util import get_icon_file_path
 
 
 class EventViewer(QDialog):
